Add basic User setup
