+++ index.html
@@ -13,8 +13,9 @@
 </head>
 
 <body>
+  
     <h1 class="display-4 text-center py-5">Tic Tac Toe</h1>
-    <h5 class="text-center d-none">Player 1 turn</h5>
+    <h5 class="text-center d-none"></h5>
     <div class="container py-5 d-none">
         <div class="row justify-content-center">
             <div class="col-2 display-3 border-bottom border-dark border-right text-center slot"></div>
@@ -40,16 +41,17 @@
     </div>
     <div class="container d-flex justify-content-center">
         <form  class="d-flex flex-column col-6">
-            <label for="player1Name">Player 1 name </label>
-            <input type="text" name="player1Name">
-            <label for="player2Name">Player 2 name </label>
-            <input type="text" name="player2Name">
-            <input type="submit" id="startBtn" class= "btn btn-success w-25 mt-2" value="Start"></input>
+            <label for="player1Name">Player 1 :</label>
+            <input type="text" name="player1Name" >
+            <label for="player2Name">Player 2 : </label>
+            <input type="text" name="player2Name"  >
+            <input type="Submit" id="startBtn" class= "btn btn-success w-25 mt-2" value="Start" ></input>
         </form>
     </div>
     <!-- Optional JavaScript -->
     <!-- jQuery first, then Popper.js, then Bootstrap JS -->
-    <script src='./assets/js/index.js' type="text/javascript"></script>
+    
+    <script src='./assets/js/index.js' type="module"></script>
     <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous">
     </script>
     <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous">
